Refactor handling of `focusTrap.onDeactivate` method
Turns out `focus-trap` was not the culprit here, but the fact that we resolved the modal _after_ calling the `deactivate` method on the focus trap.

The changes to the events in the dependency were purely coincidental, but using the newly introduced events happended to fix the issue as well as the `onPostDeactivate` happens to run a tick later than the `onDeactivate` which is enough for the `resolve` to happen.

Re #565